requirements.txt
flask

numpy

flask-restful

flask_sqlalchemy

requests